fixed some warnings
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/plugins/protein_cuda/trunk@337 8215d0b1-3b2f-4a13-8d91-4b3e73067666

+++ ProteinMovementData.cpp
@@ -566,7 +566,7 @@ bool protein::ProteinMovementData::ParsePDBAtom ( const vislib::StringA &line )
             this->tmp_atomTypes[tmpTypeTableId].Colour()[1],
             this->tmp_atomTypes[tmpTypeTableId].Colour()[2]) );
         // set atom type index to the index of the newly added atom type
-        atomTypeIdx = this->atomTypes.Count() - 1;
+        atomTypeIdx = static_cast<unsigned int>(this->atomTypes.Count() - 1);
     }
 
     // add atom data to protein atom data array
